requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-crispy-forms==1.4.0
django-toolbelt==0.0.1
django-wysiwyg-redactor==0.4.5.1
gunicorn==19.3.0
ipdb==0.8
ipython==3.1.0
Pillow==2.8.1
psycopg2==2.6
static3==0.5.1
whitenoise==1.0.6
